+++ src/mol-canvas3d/passes/draw.ts
@@ -167,7 +167,7 @@ export class DrawPass {
 
         // render opaque primitives
         if (scene.hasOpaque) {
-            renderer.renderOpaque(scene.primitives, camera, null);
+            renderer.renderOpaque(scene.primitives, camera);
         }
 
         this.depthTextureOpaque.detachFramebuffer(this.colorTarget.framebuffer, 'depth');
@@ -213,9 +213,7 @@ export class DrawPass {
 
         // render transparent volumes
         if (scene.volumes.renderables.length > 0) {
-            const target = isPostprocessingEnabled ? this.postprocessing.target : this.colorTarget;
-            target.bind();
-            renderer.renderDpoitVolume(scene.volumes, camera, this.depthTextureOpaque);
+            renderer.renderVolume(scene.volumes, camera, this.depthTextureOpaque);
         }
     }
 
@@ -227,7 +225,7 @@ export class DrawPass {
 
         // render opaque primitives
         if (scene.hasOpaque) {
-            renderer.renderOpaque(scene.primitives, camera, null);
+            renderer.renderOpaque(scene.primitives, camera);
         }
 
         if (PostprocessingPass.isTransparentDepthRequired(postprocessingProps)) {
@@ -285,7 +283,7 @@ export class DrawPass {
 
         renderer.clear(true);
         if (scene.hasOpaque) {
-            renderer.renderOpaque(scene.primitives, camera, null);
+            renderer.renderOpaque(scene.primitives, camera);
         }
 
         if (!toDrawingBuffer) {
@@ -294,7 +292,7 @@ export class DrawPass {
             if (this.depthTargetOpaque) {
                 this.depthTargetOpaque.bind();
                 renderer.clearDepth(true);
-                renderer.renderDepthOpaque(scene.primitives, camera, null);
+                renderer.renderDepthOpaque(scene.primitives, camera);
                 this.colorTarget.bind();
             }
 
@@ -320,7 +318,7 @@ export class DrawPass {
                     }
                 }
 
-                renderer.renderBlendedTransparent(scene.primitives, camera, this.depthTextureOpaque);
+                renderer.renderBlendedTransparent(scene.primitives, camera);
 
                 if (isPostprocessingEnabled) {
                     if (!this.packedDepth) {
@@ -358,7 +356,7 @@ export class DrawPass {
                 }
                 target.bind();
 
-                renderer.renderBlendedVolume(scene.volumes, camera, this.depthTextureOpaque);
+                renderer.renderVolume(scene.volumes, camera, this.depthTextureOpaque);
 
                 if (!this.packedDepth) {
                     this.depthTextureOpaque.attachFramebuffer(target.framebuffer, 'depth');
@@ -367,10 +365,10 @@ export class DrawPass {
                 }
                 target.bind();
             }
-        } else {
-            if (scene.opacityAverage < 1) {
-                renderer.renderBlendedTransparent(scene.primitives, camera, null);
-            }
+        }
+
+        if (scene.opacityAverage < 1) {
+            renderer.renderBlendedTransparent(scene.primitives, camera);
         }
     }
 
@@ -412,7 +410,7 @@ export class DrawPass {
             if (markingDepthTest && scene.markerAverage !== 1) {
                 this.marking.depthTarget.bind();
                 renderer.clear(false, true);
-                renderer.renderMarkingDepth(scene.primitives, camera, null);
+                renderer.renderMarkingDepth(scene.primitives, camera);
             }
 
             this.marking.maskTarget.bind();
@@ -479,7 +477,7 @@ export class DrawPass {
                 this.bloom.emissiveTarget.bind();
                 renderer.clear(false, true);
                 renderer.update(camera, scene);
-                renderer.renderEmissive(scene.primitives, camera, null);
+                renderer.renderEmissive(scene.primitives, camera);
             }
 
             if (!emissiveBloom || scene.emissiveAverage > 0) {