+++ src/mol-gl/shader-code.ts
@@ -325,8 +325,6 @@ const glsl300FragPrefixCommon = `
 
 #define gl_FragColor out_FragData0
 #define gl_FragDepthEXT gl_FragDepth
-
-#define depthTextureSupport
 `;
 
 function getGlsl300VertPrefix(extensions: WebGLExtensions, shaderExtensions: ShaderExtensions) {
@@ -392,6 +390,9 @@ function getGlsl300FragPrefix(gl: WebGL2RenderingContext, extensions: WebGLExten
         if (extensions.shaderTextureLod) {
             prefix.push('#define enabledShaderTextureLod');
         }
+    }
+    if (extensions.depthTexture) {
+        prefix.push('#define depthTextureSupport');
     }
     prefix.push(glsl300FragPrefixCommon);
     return prefix.join('\n') + '\n';